Move InputTypeEnum to component
