+++ src/components/Input/index.ts
@@ -1,1 +1,2 @@
 export * from './Input';
+export * from './types';
