+++ src/components/Input/types.ts
@@ -1,4 +1,28 @@
-import { InputTypeEnum, ValidationRules } from '../../../typings/commonTypes';
+import { ValidationRules } from '../../../typings/commonTypes';
+
+export enum InputTypeEnum {
+  checkbox = 'checkbox',
+  color = 'color',
+  date = 'date',
+  datetime_local = 'datetime-local',
+  email = 'email',
+  file = 'file',
+  hidden = 'hidden',
+  image = 'image',
+  month = 'month',
+  number = 'number',
+  password = 'password',
+  radio = 'radio',
+  range = 'range',
+  reset = 'reset',
+  search = 'search',
+  submit = 'submit',
+  tel = 'tel',
+  text = 'text',
+  time = 'time',
+  url = 'url',
+  week = 'week',
+}
 
 export type InputProps = {
   inputHandler?(value: string): void;
